requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
cffi==1.16.0
Django==5.0.4
django-autoslug==1.9.9
django-debug-toolbar==4.3.0
django-filter==24.2
django-sortedm2m==3.1.1
djangorestframework==3.15.1
Markdown==3.6
pillow==10.3.0
psycopg2-binary==2.9.9
pycparser==2.22
pytils==0.4.1
sqlparse==0.5.0
